+++ pytest_inmanta/plugin.py
@@ -18,6 +18,7 @@
 import collections
 import importlib
 import inspect
+import itertools
 import json
 import logging
 import math
@@ -34,7 +35,7 @@ from itertools import chain
 from pathlib import Path
 from textwrap import dedent
 from types import FunctionType, ModuleType
-from typing import Dict, Iterator, List, Optional, Set, Tuple
+from typing import Dict, Iterator, List, Optional, Sequence, Set, Tuple
 
 import pydantic
 import pytest
@@ -55,6 +56,7 @@ from inmanta.execute.proxy import DynamicProxy
 from inmanta.export import Exporter, ResourceDict, cfg_env
 from inmanta.protocol import json_encode
 from inmanta.resources import Resource
+from pytest_inmanta.core import SUPPORTS_PROJECT_PIP_INDEX
 
 if typing.TYPE_CHECKING:
     # Local type stub for mypy that works with both pytest < 7 and pytest >=7
@@ -67,6 +69,9 @@ if typing.TYPE_CHECKING:
             ...
 
 
+if SUPPORTS_PROJECT_PIP_INDEX:
+    from inmanta.module import ProjectPipConfig
+
 from .handler import DATA
 from .parameters import (
     inm_install_mode,
@@ -75,6 +80,7 @@ from .parameters import (
     inm_no_load_plugins,
     inm_no_strict_deps_check,
     inm_venv,
+    pip_index_urls,
 )
 from .test_parameter import ParameterNotSetException, TestParameterRegistry
 
@@ -218,6 +224,12 @@ def get_project_repos(repo_options: typing.Sequence[str]) -> typing.Sequence[obj
             # there might be only one part or part might be just "https"
             except (IndexError, pydantic.ValidationError):
                 repo_info = module.ModuleRepoInfo(url=repo_str)
+            if SUPPORTS_PROJECT_PIP_INDEX:
+                if repo_info.type == module.ModuleRepoType.package:
+                    LOGGER.warning(
+                        "Setting a package source through the --module-repo <index_url> with type `package` "
+                        "is now deprecated in favour of the --pip-index-urls <index_url> option.`"
+                    )
             return json.loads(repo_info.json())
 
     return [parse_repo(repo) for repo in repo_options]
@@ -261,19 +273,47 @@ def project_metadata(request: pytest.FixtureRequest) -> module.ProjectMetadata:
         )
     )
 
+    index_urls: Sequence[str] = pip_index_urls.resolve(request.config)
     modulepath = ["libs"]
     in_place = inm_mod_in_place.resolve(request.config)
     if in_place:
         modulepath.append(str(Path(CURDIR).parent))
 
-    return module.ProjectMetadata(
-        name="testcase",
-        description="Project for testcase",
-        repo=repos,
-        modulepath=modulepath,
-        downloadpath="libs",
-        install_mode=inm_install_mode.resolve(request.config).value,
-    )
+    if SUPPORTS_PROJECT_PIP_INDEX:
+        # On newer versions of core we set the pip.index_url of the project.yml file
+        repos_urls: List[str] = [
+            repo["url"]
+            for repo in repos
+            if repo["type"] == module.ModuleRepoType.package
+        ]
+        pip_config: ProjectPipConfig = ProjectPipConfig(
+            # This ensures no duplicates are returned and insertion order is preserved.
+            # i.e. the left-most index will be passed to pip as --index-url and the others as --extra-index-url
+            index_urls=list(
+                {value: None for value in itertools.chain(index_urls, repos_urls)}
+            )
+        )
+        return module.ProjectMetadata(
+            name="testcase",
+            description="Project for testcase",
+            repo=repos,
+            modulepath=modulepath,
+            downloadpath="libs",
+            install_mode=inm_install_mode.resolve(request.config).value,
+            pip=pip_config,
+        )
+    else:
+        v2_source_repos = [
+            {"url": index_url, "type": "package"} for index_url in index_urls
+        ]
+        return module.ProjectMetadata(
+            name="testcase",
+            description="Project for testcase",
+            repo=list(repos) + v2_source_repos,
+            modulepath=modulepath,
+            downloadpath="libs",
+            install_mode=inm_install_mode.resolve(request.config).value,
+        )
 
 
 @pytest.fixture(scope="session")